Primitive Physics Integration
This is the first implementation of physics within the game. The next
step is to make the sprite actually control, and then make a script that
randomly changes the way the character is shifted. Along those lines, we
need to make it so that the sprite rotates based on the direction in
which the gravity initially spawns.