Implemented new Hitboxes and transparent sprites
I may have changed the physics settings of gravity to -30 based on a
tutorial that I am using in order to get animations working.
Changed size of hitbox to match the floor
Updated sprite to be transparent, (still needs to be un blurred)
BUGS: Stage moves horizontally as soon as the polygon hitbox touches the
floor hitbox.
Next step is to add character controllers and Sprite animation.